Moving reducers to another file
