+++ src/protspace/utils/reducers.py
@@ -1,0 +1,354 @@
+import inspect
+from abc import ABC, abstractmethod
+from dataclasses import dataclass, field, fields
+from typing import Any, Dict, List, get_args, get_type_hints, Literal
+
+import logging
+import numpy as np
+from sklearn.decomposition import PCA
+from sklearn.manifold import MDS, TSNE
+from umap import UMAP
+from pacmap import PaCMAP, LocalMAP
+
+
+# Method names constants
+PCA_NAME = "pca"
+TSNE_NAME = "tsne"
+UMAP_NAME = "umap"
+PACMAP_NAME = "pacmap"
+MDS_NAME = "mds"
+LOCALMAP_NAME = "localmap"
+
+ALL_METHODS = [PCA_NAME, TSNE_NAME, UMAP_NAME, PACMAP_NAME, MDS_NAME, LOCALMAP_NAME]
+
+# Metric types
+METRIC_TYPES = Literal["euclidean", "cosine"]
+
+logging.basicConfig(format="%(levelname)s: %(message)s")
+logger = logging.getLogger(__name__)
+
+@dataclass(frozen=True)
+class DimensionReductionConfig:
+    """Configuration for dimension reduction methods.
+
+    Parameters:
+        n_components: Number of dimensions in reduced space (2 or 3)
+        n_neighbors: Number of neighbors for manifold learning (>0)
+        metric: Distance metric to use
+        precomputed: Whether distances are precomputed
+        min_dist: Minimum distance for UMAP (0-1)
+        perplexity: Perplexity for t-SNE (5-50)
+        learning_rate: Learning rate for t-SNE optimization (>0)
+        mn_ratio: Ratio for PaCMAP (0-1)
+        fp_ratio: Ratio for PaCMAP (>0)
+        n_init: Number of initializations for MDS (>0)
+        max_iter: Maximum iterations (>0)
+        eps: Convergence tolerance (>0)
+    """
+
+    n_components: int = field(default=2, metadata={"allowed": [2, 3]})
+    n_neighbors: int = field(default=15, metadata={"gt": 0})
+    metric: METRIC_TYPES = field(
+        default="euclidean", metadata={"allowed": [m for m in get_args(METRIC_TYPES)]}
+    )
+    precomputed: bool = field(default=False)
+    min_dist: float = field(default=0.1, metadata={"gte": 0, "lte": 1})
+    perplexity: int = field(default=30, metadata={"gte": 5, "lte": 50})
+    learning_rate: int = field(default=200, metadata={"gt": 0})
+    mn_ratio: float = field(default=0.5, metadata={"gte": 0, "lte": 1})
+    fp_ratio: float = field(default=2.0, metadata={"gt": 0})
+    n_init: int = field(default=4, metadata={"gt": 0})
+    max_iter: int = field(default=300, metadata={"gt": 0})
+    eps: float = field(default=1e-3, metadata={"gt": 0})
+
+    def __post_init__(self):
+        """Validate configuration parameters."""
+        for data_field in fields(self):
+            value = getattr(self, data_field.name)
+            metadata = data_field.metadata
+
+            if "allowed" in metadata:
+                if value not in metadata["allowed"]:
+                    raise ValueError(
+                        f"{data_field.name} must be one of {metadata['allowed']}"
+                    )
+
+            if "gt" in metadata:
+                if value <= metadata["gt"]:
+                    raise ValueError(
+                        f"{data_field.name} must be greater than {metadata['gt']}"
+                    )
+
+            if "lt" in metadata:
+                if value >= metadata["lt"]:
+                    raise ValueError(
+                        f"{data_field.name} must be less than {metadata['lt']}"
+                    )
+
+            if "gte" in metadata:
+                if value < metadata["gte"]:
+                    raise ValueError(
+                        f"{data_field.name} must be greater than or equal to {metadata['gte']}"
+                    )
+
+            if "lte" in metadata:
+                if value > metadata["lte"]:
+                    raise ValueError(
+                        f"{data_field.name} must be less than or equal to {metadata['lte']}"
+                    )
+
+    def parameters_by_method(self, method: str) -> List[Dict[str, Any]]:
+        method_map = {
+            TSNE_NAME: TSNE,
+            PCA_NAME: PCA,
+            UMAP_NAME: UMAP,
+            PACMAP_NAME: PaCMAP,
+            MDS_NAME: MDS,
+            LOCALMAP_NAME: LocalMAP,
+        }
+
+        if method not in method_map:
+            return []
+
+        def _get_parameter_desc_from_docstring(parameter: str, docstring: str) -> str:
+            large_splits = []
+            possible_split_variants = [
+                f"{parameter} : ",
+                f"{parameter}: ",
+                f"{parameter}:",
+                parameter,
+            ]
+            for split_variant in possible_split_variants:
+                if split_variant in docstring:
+                    large_splits = docstring.split(split_variant)
+                    break
+            if len(large_splits) == 0:
+                return ""
+            large_split = large_splits[0] if len(large_splits) == 1 else large_splits[1]
+            param_split = (
+                large_split.split("\n\n")[0]
+                if "\n" in large_split
+                else large_split.split("\n")[0]
+            )
+            param_split_cleaned = (
+                param_split.replace("\n\n", "")
+                .replace("\t", "")
+                .replace("  ", " ")
+                .replace("   ", " ")
+                .strip()
+            )
+            return param_split_cleaned
+
+        type_hints = get_type_hints(self.__class__)
+
+        try:
+            method_function = method_map[method]
+            method_signature = inspect.signature(method_function)
+            docstring = inspect.getdoc(method_function)
+            method_parameters = list(method_signature.parameters.keys())
+            # Create a dictionary of lowercase attribute names to their original names
+            lowercase_fields = {
+                data_field.name.lower(): data_field for data_field in fields(self)
+            }
+            result = []
+            for param in method_parameters:
+                # Exclude parameters not relevant for certain methods
+                if method == MDS_NAME and param == "metric":
+                    continue
+                if method == UMAP_NAME and param == "learning_rate":
+                    continue
+                if method == LOCALMAP_NAME and param == "metric":
+                    continue
+
+                if param.lower() in lowercase_fields:
+                    data_field = lowercase_fields[param.lower()]
+                    field_type_hint = type_hints.get(data_field.name, Any)
+                    field_type_name = getattr(
+                        field_type_hint, "__name__", str(field_type_hint)
+                    )
+                    if hasattr(
+                        field_type_hint, "__args__"
+                    ):  # Handle Literal, Union etc.
+                        field_type_name = str(field_type_hint).replace("typing.", "")
+
+                    doc_desc = _get_parameter_desc_from_docstring(
+                        parameter=param, docstring=docstring
+                    )
+                    description = (
+                        doc_desc
+                        if doc_desc
+                        else f"{data_field.name}: Config parameter. Default: {data_field.default}"
+                    )
+
+                    result.append(
+                        {
+                            "name": param.lower(),
+                            "default": data_field.default,
+                            "description": description,
+                            "constraints": {
+                                "type": field_type_name,
+                                **data_field.metadata,
+                            },
+                        }
+                    )
+            return result
+        except Exception as e:
+            print(e)
+            return []
+
+
+class DimensionReducer(ABC):
+    """Abstract base class for dimension reduction methods."""
+
+    def __init__(self, config: DimensionReductionConfig):
+        self.config = config
+
+    @abstractmethod
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        """Transform data to lower dimensions."""
+        pass
+
+    @abstractmethod
+    def get_params(self) -> Dict[str, Any]:
+        """Get parameters used for the reduction."""
+        pass
+
+
+class PCAReducer(DimensionReducer):
+    """Principal Component Analysis reduction, preferring ARPACK solver."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        solver = "arpack"
+        n_samples, n_features = data.shape
+        k = self.config.n_components
+
+        # ARPACK requires n_components < min(shape), fallback to full SVD otherwise
+        if k >= min(n_samples, n_features):
+            logger.warning(
+                f"PCA: n_components ({k}) >= min(shape) ({min(n_samples, n_features)}). "
+                f"'arpack' solver unavailable, falling back to 'full' solver."
+            )
+            solver = "full"
+
+        pca = PCA(n_components=k, svd_solver=solver)
+        try:
+            result = pca.fit_transform(data)
+            self.explained_variance = pca.explained_variance_ratio_.tolist()
+            self.used_solver = solver  # Store the solver that was actually used
+            return result
+        except Exception as e:
+            logger.error(f"PCA failed using '{solver}' solver: {e}")
+            raise
+
+    def get_params(self) -> Dict[str, Any]:
+        """Get parameters used for the reduction."""
+        params = {
+            "n_components": self.config.n_components,
+            # Report the solver used, default to 'arpack' if not set yet
+            "svd_solver": getattr(self, "used_solver", "arpack"),
+        }
+        if hasattr(self, "explained_variance"):
+            params["explained_variance_ratio"] = self.explained_variance
+        return params
+
+
+class TSNEReducer(DimensionReducer):
+    """t-SNE (t-Distributed Stochastic Neighbor Embedding) reduction."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        return TSNE(
+            n_components=self.config.n_components,
+            perplexity=self.config.perplexity,
+            learning_rate=self.config.learning_rate,
+            metric=self.config.metric,
+        ).fit_transform(data)
+
+    def get_params(self) -> Dict[str, Any]:
+        return {
+            "n_components": self.config.n_components,
+            "perplexity": self.config.perplexity,
+            "learning_rate": self.config.learning_rate,
+            "metric": self.config.metric,
+        }
+
+
+class UMAPReducer(DimensionReducer):
+    """UMAP (Uniform Manifold Approximation and Projection) reduction."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        return UMAP(
+            n_components=self.config.n_components,
+            n_neighbors=self.config.n_neighbors,
+            min_dist=self.config.min_dist,
+            metric=self.config.metric,
+        ).fit_transform(data)
+
+    def get_params(self) -> Dict[str, Any]:
+        return {
+            "n_components": self.config.n_components,
+            "n_neighbors": self.config.n_neighbors,
+            "min_dist": self.config.min_dist,
+            "metric": self.config.metric,
+        }
+
+
+class PaCMAPReducer(DimensionReducer):
+    """PaCMAP (Pairwise Controlled Manifold Approximation) reduction."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        return PaCMAP(
+            n_components=self.config.n_components,
+            n_neighbors=self.config.n_neighbors,
+            MN_ratio=self.config.mn_ratio,
+            FP_ratio=self.config.fp_ratio,
+        ).fit_transform(data)
+
+    def get_params(self) -> Dict[str, Any]:
+        return {
+            "n_components": self.config.n_components,
+            "n_neighbors": self.config.n_neighbors,
+            "MN_ratio": self.config.mn_ratio,
+            "FP_ratio": self.config.fp_ratio,
+        }
+
+
+class LocalMAPReducer(DimensionReducer):
+    """LocalMAP (Local Manifold Approximation) reduction."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        return LocalMAP(
+            n_components=self.config.n_components,
+            n_neighbors=self.config.n_neighbors,
+            MN_ratio=self.config.mn_ratio,
+            FP_ratio=self.config.fp_ratio,
+        ).fit_transform(data, init="pca")
+
+    def get_params(self) -> Dict[str, Any]:
+        return {
+            "n_components": self.config.n_components,
+            "n_neighbors": self.config.n_neighbors,
+            "MN_ratio": self.config.mn_ratio,
+            "FP_ratio": self.config.fp_ratio,
+        }
+
+
+class MDSReducer(DimensionReducer):
+    """Multidimensional Scaling reduction."""
+
+    def fit_transform(self, data: np.ndarray) -> np.ndarray:
+        return MDS(
+            n_components=self.config.n_components,
+            metric=self.config.precomputed,
+            n_init=self.config.n_init,
+            max_iter=self.config.max_iter,
+            eps=self.config.eps,
+            dissimilarity=("precomputed" if self.config.precomputed else "euclidean"),
+        ).fit_transform(data)
+
+    def get_params(self) -> Dict[str, Any]:
+        return {
+            "n_components": self.config.n_components,
+            "n_init": self.config.n_init,
+            "max_iter": self.config.max_iter,
+            "eps": self.config.eps,
+        }
